requirements_2.txt
# ============================================
# PHASE 1: Data Collection & Foundation
# ============================================

# Snowflake & Database
snowflake-connector-python==3.7.0
snowflake-sqlalchemy==1.5.1

# dbt for transformations
dbt-core==1.7.7
dbt-snowflake==1.7.1

# Data manipulation
pandas>=2.3.3
numpy==1.26.3

# Configuration
python-dotenv>=1.2.1

# HTTP requests (for APIs)
requests==2.31.0

# Date/time utilities
python-dateutil==2.8.2

# ============================================
# AIRFLOW: Workflow Orchestration
# ============================================

# Core Airflow
apache-airflow==2.8.1

# Airflow Providers (only what we need)
apache-airflow-providers-snowflake==6.4.0
apache-airflow-providers-common-sql==1.10.0

# Database for Airflow metadata (SQLite for dev, Postgres for prod)
# SQLite comes with Python, but for better performance:
# psycopg2-binary==2.9.9  # Uncomment if using PostgreSQL

pysnowflake>=0.1.3
httpx>=0.28.1